+++ cmd/slackdump/internal/man/chunk.go
@@ -1,0 +1,16 @@
+package man
+
+import (
+	_ "embed"
+
+	"github.com/rusq/slackdump/v2/cmd/slackdump/internal/golang/base"
+)
+
+//go:embed assets/chunk.md
+var mdChunk string
+
+var Chunk = &base.Command{
+	UsageLine: "slackdump chunk",
+	Short:     "chunk file format specification",
+	Long:      mdChunk,
+}